# openapi.yaml
openapi: 3.1.0
info:
  title: Payment Service
  description: 'Production-ready Digital Wallet payment API. Provides a complete payment processing API including checkouts,
    refunds, customer management, and receipt generation.


    ## Authentication

    All API endpoints (except health checks and auth) require an `X-API-Key` header for authentication. Click the **Authorize**
    button (🔒) above to enter your API key, and it will be included in all requests.


    - **Admin key**: Set via `ADMIN_API_KEY` env var — used to manage tenant API keys

    - **Tenant keys**: Created via `POST /api/v1/admin/api-keys` — used for payment operations


    ## Idempotency

    POST endpoints support the `Idempotency-Key` header to prevent duplicate operations.


    ## Rate Limiting

    API requests are rate-limited per API key. Each tenant key can have custom rate limits. Check the `X-RateLimit-*` response
    headers for current limits.'
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development server
paths:
...
  /api/v1/checkout/sessions:
    post:
      tags:
      - Checkout
      summary: Create a custom checkout session
      description: 'Create a hosted checkout session. Returns a `checkout_url` — redirect the end-user''s browser to that
        URL to collect their custom Digital Wallet authorization.


        **This is the main integration point for client services.** The client only needs to call this endpoint, redirect
        the user to the local UI, and wait for the success callback redirect.'
      operationId: create_checkout_session_api_v1_checkout_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutSessionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/checkout/sessions/{session_id}/authorize:
    post:
      tags:
      - Checkout
      summary: Authorize a checkout session via Password
      operationId: authorize_checkout_session_api_v1_checkout_sessions__session_id__authorize_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutAuthorizeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutAuthorizeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/checkout/sessions/{session_id}:
    get:
      tags:
      - Checkout
      summary: Get checkout session status
      description: 'Check the status of a checkout session. Use this to verify payment completion after the user is redirected
        back.


        **Statuses**: `open` (waiting for payment), `complete` (paid), `expired` (timed out, ~24h).'
      operationId: get_checkout_session_api_v1_checkout_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/customers:
    post:
      tags:
      - Customers
      summary: Create a customer
      description: Create a new customer in the system.
      operationId: create_customer_api_v1_customers_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Customers
      summary: List customers
      description: Retrieve a paginated list of customers.
      operationId: list_customers_api_v1_customers_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Limit
        description: Items per page
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of items to skip
          default: 0
          title: Offset
        description: Number of items to skip
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by email (partial match)
          title: Email
        description: Filter by email (partial match)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/customers/me/transactions:
    get:
      tags:
      - Customers
      summary: Get my transactions
      description: Retrieve the payment history for the currently authenticated customer.
      operationId: get_my_transactions_api_v1_customers_me_transactions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Limit
        description: Items per page
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of items to skip
          default: 0
          title: Offset
        description: Number of items to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/customers/{customer_id}:
    get:
      tags:
      - Customers
      summary: Get customer details
      description: Retrieve details of a specific customer by ID.
      operationId: get_customer_api_v1_customers__customer_id__get
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Customers
      summary: Update a customer
      description: Update customer details. Only provided fields will be changed.
      operationId: update_customer_api_v1_customers__customer_id__put
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Customers
      summary: Delete a customer
      description: Soft-delete a customer.
      operationId: delete_customer_api_v1_customers__customer_id__delete
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CheckoutAuthorizeRequest:
      properties:
        password:
          type: string
          title: Password
          description: Customer's wallet password to authorize the transaction
      type: object
      required:
      - password
      title: CheckoutAuthorizeRequest
    CheckoutAuthorizeResponse:
      properties:
        status:
          type: string
          title: Status
          description: Authorization status (e.g. 'succeeded')
        payment_id:
          type: string
          format: uuid
          title: Payment Id
          description: ID of the completed internal payment
        success_url:
          type: string
          title: Success Url
          description: URL to redirect the customer to upon success
      type: object
      required:
      - status
      - payment_id
      - success_url
      title: CheckoutAuthorizeResponse
    CheckoutLineItem:
      properties:
        name:
          type: string
          maxLength: 200
          title: Name
          description: Product/item name shown on the checkout page
        quantity:
          type: integer
          minimum: 1.0
          title: Quantity
          description: Number of units
          default: 1
        price:
          type: integer
          exclusiveMinimum: 0.0
          title: Price
          description: Unit price in smallest currency unit (e.g., cents)
      type: object
      required:
      - name
      - price
      title: CheckoutLineItem
      description: A single line item in the checkout.
    CheckoutSessionCreate:
      properties:
        line_items:
          items:
            $ref: '#/components/schemas/CheckoutLineItem'
          type: array
          minItems: 1
          title: Line Items
          description: Items being purchased (displayed on the checkout page)
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          description: ISO 4217 currency code
          default: usd
        success_url:
          type: string
          title: Success Url
          description: URL to redirect the customer after successful payment.
        cancel_url:
          type: string
          title: Cancel Url
          description: URL to redirect the customer if they cancel the checkout.
        customer_email:
          type: string
          title: Customer Email
          description: Email of the customer. If they don't exist in the Payment Service, a new identity will be created automatically.
        customer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Name
          description: Name of the customer (used if creating a new identity)
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Arbitrary key-value metadata passed to the payment
      type: object
      required:
      - line_items
      - success_url
      - cancel_url
      - customer_email
      title: CheckoutSessionCreate
      description: 'Request body to create a hosted checkout session.


        The client only needs to send the payment details and redirect URLs.

        The response includes a checkout_url — redirect the end-user there.'
      examples:
      - cancel_url: https://mytickets.com/order/123/cancel
        currency: usd
        customer_email: alice@example.com
        customer_name: Alice Smith
        line_items:
        - name: VIP Concert Ticket
          price: 5000
          quantity: 2
        metadata:
          event_id: evt_456
          order_id: order_123
        success_url: https://mytickets.com/order/123/success
    CheckoutSessionResponse:
      properties:
        session_id:
          type: string
          title: Session Id
          description: Checkout Session ID
        line_items:
          anyOf:
          - items:
              $ref: '#/components/schemas/CheckoutLineItem'
            type: array
          - type: 'null'
          title: Line Items
          description: Items to display in the UI
        checkout_url:
          type: string
          title: Checkout Url
          description: URL to redirect the customer to for payment
        status:
          type: string
          title: Status
          description: 'Session status: open, complete, or expired'
        success_url:
          type: string
          title: Success Url
          description: URL to redirect on success
        cancel_url:
          type: string
          title: Cancel Url
          description: URL to redirect on cancel
        expires_at:
          type: string
          format: date-time
          title: Expires At
          description: When this checkout session expires
        payment_status:
          type: string
          title: Payment Status
          description: 'Payment status: unpaid, paid, or no_payment_required'
        amount_total:
          type: integer
          title: Amount Total
          description: Total amount in smallest currency unit
        currency:
          type: string
          title: Currency
          description: ISO 4217 currency code
        customer_name:
          type: string
          title: Customer Name
          description: Customer name
        customer_email:
          type: string
          title: Customer Email
          description: Customer email
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Metadata
      type: object
      required:
      - session_id
      - checkout_url
      - status
      - success_url
      - cancel_url
      - expires_at
      - payment_status
      - amount_total
      - currency
      - customer_name
      - customer_email
      title: CheckoutSessionResponse
      description: 'Response after creating a checkout session.


        Redirect the end-user''s browser to `checkout_url` to complete payment.'
    CustomerCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
          description: Customer email address
        name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Name
          description: Customer full name
        phone:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Phone
          description: Customer phone number
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Arbitrary key-value metadata
      type: object
      required:
      - email
      title: CustomerCreate
      description: Request body to create a customer.
      examples:
      - email: john.doe@example.com
        metadata:
          external_id: user_123
        name: John Doe
        phone: '+1234567890'
    CustomerListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CustomerResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: CustomerListResponse
      description: Paginated list of customers.
...
    CustomerResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - created_at
      - updated_at
      title: CustomerResponse
      description: Customer response body.
    CustomerUpdate:
      properties:
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
          description: Customer email address
        name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Name
          description: Customer full name
        phone:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Phone
          description: Customer phone number
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Arbitrary key-value metadata
      type: object
      title: CustomerUpdate
      description: Request body to update a customer.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaymentListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PaymentResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: PaymentListResponse
      description: Paginated list of payments.
    PaymentResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        customer_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Customer Id
        amount:
          type: integer
          title: Amount
        currency:
          type: string
          title: Currency
        status:
          type: string
          title: Status
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
        amount_refunded:
          type: integer
          title: Amount Refunded
          default: 0
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - amount
      - currency
      - status
      - created_at
      - updated_at
      title: PaymentResponse
      description: Payment response body.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Enter your API key. Use the **admin key** (from ADMIN_API_KEY env var) for `/api/v1/admin/*` endpoints,
        or a **tenant key** (ps_live_...) for payment/refund/customer operations.
...
security:
- ApiKeyAuth: []